修复 openwrt 编译 r2s失败，修改 sft1200 适配新缓存步骤

+++ build/sft1200/diy-after-compile.sh
@@ -1,3 +1,6 @@
-rm -f $(dirname $(find openwrt/bin/targets/ -type f -name sha256sums ))/openwrt-*-root.squashfs
-
+# sft1200 不考虑做 slim 和 full，默认会生成 imagebuilder，这里干掉
+if [ -n "${GITHUB_RUN_NUMBER}" ];then
+    rm -f $(dirname $(find openwrt/bin/targets/ -type f -name sha256sums ))/openwrt-*-root.squashfs
+    #rm -f $(dirname $(find openwrt/bin/targets/ -type f -name sha256sums ))/*-imagebuilder-*
+fi
 true
